file organizer 5.0

+++ File_Management/instaRun.py
@@ -1,9 +1,29 @@
 import os
 import shutil
 
+download_path = "C:\\Users\\Denis\\Desktop\\Downloads"
 
 
-download_path = "C:\\Users\\Denis\\Downloads\\"
+with open('C:\\Users\\Denis\\Downloads\\.Folders\\Extensions_file.txt', 'r') as file:
+    images = file.readline()
+    text = file.readline()
+    videos = file.readline()
+    sounds = file.readline()
+    applications = file.readline()
+    codes = file.readline()
+    zip_files = file.readline()
+
+
+extensions = [images, text, videos, sounds, applications, codes, zip_files]
+
+
+for i, j in enumerate(extensions):
+
+    j = j.split(', ')
+    j[0] = j[0].split('= ')[1]
+    j[-1] = j[-1].split('\n')[0]
+    extensions[i] = j
+
 
 s = os.chdir(download_path)
 
@@ -11,63 +31,43 @@ current = os.getcwd()
 
 files = os.listdir(current)
 
-images = [".jpeg", ".png", ".jpg", ".gif", ".jfif"]  # extensions for images
-text = [".doc", ".txt", ".pdf", ".xlsx", ".docx", ".xls", ".rtf", ".md"]  # extensions for text files
-videos = [".mp4", ".mkv"]  # extensions for videos
-sounds = [".mp3", ".wav", ".m4a"]  # extensions for sounds
-applications = [".exe", ".lnk", ".msi"]  # extensions for applications
-codes = [".c", ".java", ".py", ".cpp", ".js", ".html", ".css", ".php"]  # extensions for codes
-zip_files = [".zip", ".rar"]  # extension for zip files
 folders = [".Folders", ".Installers", ".Music", ".Other", ".Random Code",
            ".Saved Pictures", ".Saved Videos", ".Text",
            ".Zip Files"]  # all the folders that are excepted from organizing
 
+for folder in folders:
+    if folder not in files:
+        os.mkdir(os.path.join(download_path, folder))
+
+dict = {".Saved Pictures": extensions[0], ".Text": extensions[1],
+        ".Saved Videos": extensions[2], ".Music": extensions[3],
+        ".Installers": extensions[4], ".Random Code": extensions[5],
+        ".Zip Files": extensions[6]}
+
+keys_list = list(dict.keys())
 
 print("Sorting the files...")
 
+
 for file in files:
     destination = ""
-    for ex in images:
-        if file.endswith(ex):
-            destination = os.path.join(download_path, ".Saved Pictures")
-            shutil.move(file, destination)
-            break
 
-    for ex in text:
-        if file.endswith(ex):
-            destination = os.path.join(download_path, ".Text")
-            shutil.move(file, destination)
-            break
+    for i in range(len(keys_list)):
+        for ex in dict[keys_list[i]]:
+            if file.endswith(ex):
+                destination = os.path.join(download_path, keys_list[i])
 
-    for ex in sounds:
-        if file.endswith(ex):
-            destination = os.path.join(download_path, ".Music")
-            shutil.move(file, destination)
-            break
+                if os.path.isfile(os.path.join(destination, file)):
+                    try:
+                        shutil.move(file, os.path.join(download_path, ".Other\\.Duplicates"))
+                    except shutil.Error:
+                        new_path_file = os.path.join(download_path, ".Other\\.Duplicates")
+                        os.rename(file, os.path.join(new_path_file, f"DUP {file}"))
+                else:
+                    shutil.move(file, destination)
 
-    for ex in videos:
-        if file.endswith(ex):
-            destination =  os.path.join(download_path, ".Saved Videos")
-            shutil.move(file, destination)
-            break
+                break
 
-    for ex in applications:
-        if file.endswith(ex):
-            destination = os.path.join(download_path, ".Installers")
-            shutil.move(file, destination)
-            break
-
-    for ex in codes:
-        if file.endswith(ex):
-            destination = os.path.join(download_path, ".Random Code")
-            shutil.move(file, destination)
-            break
-
-    for ex in zip_files:
-        if file.endswith(ex):
-            destination = os.path.join(download_path, ".Zip Files")
-            shutil.move(file, destination)
-            break
     if os.path.isfile(file):
         if destination == "":
             shutil.move(file, os.path.join(download_path, ".Other"))
